requirements.txt
pytorch
torchvision
numpy
random
copy
opencv-python
scipy
argparse
logging
kornia
